Move pyhdf to conda environment file
Install of pyhdf on macos was failing for me becaues I didn't have hdf
installed. For now I have moved the `pyhdf` requirement to the conda
environment file so that it is installed with conda instead. Eventually
we should instead have a "optional extra" part of `cloudmetrics`, maybe
called `cloudmetrics[extra]` which will require extra system packages to
already be installed (and then mention this in the readme).