Add setup.py to enable installation with pip (#26)
Install of pyhdf on macos was failing on macos because hdf
is not installed by default. For now `pyhdf` requirement is moved to the conda
environment file so that it is installed with conda instead. Eventually
we should instead have a "optional extra" part of `cloudmetrics`, maybe
called `cloudmetrics[extra]` which will require extra system packages to
already be installed (and then mention this in the readme).

Co-authored-by: martinjanssens <martinjanssens94@hotmail.com>